copyCat/requirements.txt
   Flask
   nltk
   spacy
   sentence-transformers
   python-Levenshtein
   matplotlib